# backend/src/main/resources/static/api/api.yaml
openapi: 3.0.3
info:
  title: Ya.Stream - OpenAPI 3.0
  description: |-
    TODO Description
    
    Some useful links:
    - [Ya.Stream repository](https://github.com/HaskSy/Ya.Stream)
  version: 1.0.0
servers:
  - url: https://music.gvsem.com/api/
tags:
  - name: listen
    description: SSE listener of streamer's actions
  - name: stream
    description: Streamer actions
paths:

  /listen/{yandex_login}:
    get:
      operationId: listenToUserId
      tags:
        - listen
      summary: SSE subscription to audio of specified user
      description: NO AUTHENTICATION IS REQUIRED FOR THIS METHOD
      parameters:
        - in: path
          name: yandex_login
          required: true
          schema:
            type: string
          description: yandex_login
      responses:
        '200':
          description: Successful operation
          content:
            text/event-stream:
              schema:
                $ref: '#/components/schemas/ListenEvent'
        '404':
          description: User does not exist or does not stream right now

  /stream:
    get:
      security:
        - BearerAuth : []
      parameters:
        - in: query
          name: event
          required: true
          schema:
            type: string
            enum: [play, stop, goto]
        - in: query
          name: track_id
          description: track id in Yandex.Music for play, stop, goto
          required: false
          schema:
            type: string
        - in: query
          name: position
          description: current player position for event for play, stop, goto
          required: false
          schema:
            type: string
      tags:
        - stream
      summary: Send recent user's action while interacting with music player
      responses:
        '201':
          description: Successful operation
          
components:
  schemas:

    ListenEvent:
      format: event-stream
      type: object
      properties:
        event:
          type: string
          nullable: false
          enum: [play, stop, next, prev, goto]
        track_id:
          type: string
          nullable: false
        position:
          type: string

  securitySchemes:
    BearerAuth:
      type: apiKey
      in: header
      name: Authorization

#security:
#  - BearerAuth: []